+++ task/z.ts
@@ -1,9 +1,7 @@
 import c2a from '../src/index.js'
 
 const main = async () => {
-  await c2a('./script/z', {
-    ast: true,
-    metadata: false,
+  await c2a('./script/z/index.coffee', {
     salt: 'ahk',
   })
 }